Various fixes for slider motion and handle design change
A collection of fixes for slider motion.
One fix of making sure that the same maths is used for conversion from min/max to center/width.
Multiple fixes for making sure event propagation did not cause undesired additional changes, mostly around the inclusion of the __suppress_updates flag which limits the updates to only a single event driven update.This removed most of the 'wobbly' effect where events triggered as part of the partial update caused other side-effects.

Also changed the handle style to be a trial of a bookend style handle to see whether that works better.